+++ tests/dataflow/test_simulator.py
@@ -3,7 +3,6 @@
 
 from allo.ir.types import float32
 import allo.dataflow as df
-import allo.backend.hls as hls
 import numpy as np
 
 M, N, K = 32, 32, 32
@@ -26,11 +25,9 @@ def test_tiled_gemm():
     A = np.random.rand(M, K).astype(np.float32)
     B = np.random.rand(K, N).astype(np.float32)
     C = np.zeros((M, N), dtype=np.float32)
-    mod = df.build(top)
-    if hls.is_available("vitis_hls"):
-        mod(A, B, C)
-        np.testing.assert_allclose(C, np.dot(A, B), rtol=1e-5, atol=1e-5)
-        print("Success!")
+    mod = df.build(top, target="simulator")
+    mod(A, B, C)
+    np.testing.assert_allclose(C, np.dot(A, B), rtol=1e-5)
 
 
 if __name__ == "__main__":
